Refactor some common code out of salesforcedx-apex-debugger into salesforcedx-utils-vscode
