+++ packages/salesforcedx-apex-debugger/src/commands/baseDebuggerCommand.ts
@@ -5,7 +5,7 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import { BaseCommand } from './baseCommand';
+import { BaseCommand } from '@salesforce/salesforcedx-utils-vscode/out/src/requestService';
 import { DebuggerRequest } from './protocol';
 
 export class BaseDebuggerCommand extends BaseCommand {
