+++ packages/salesforcedx-apex-debugger/src/commands/index.ts
@@ -5,13 +5,9 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-export { BaseCommand } from './baseCommand';
-export { ForceConfigGet } from './forceConfigGet';
-export { ForceOrgDisplay, OrgInfo } from './forceOrgDisplay';
 export { FrameCommand } from './frameCommand';
 export * from './protocol';
 export { ReferencesCommand } from './referencesCommand';
-export { RequestService, RestHttpMethodEnum } from './requestService';
 export { RunCommand } from './runCommand';
 export {
   StepIntoCommand,
