+++ packages/salesforcedx-apex-debugger/src/core/streamingService.ts
@@ -5,7 +5,7 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import { RequestService } from '../commands/requestService';
+import { RequestService } from '@salesforce/salesforcedx-utils-vscode/out/src/requestService';
 import {
   ApexDebuggerEventType,
   StreamingClient,
