+++ packages/salesforcedx-apex-debugger/test/unit/adapter/apexDebug.test.ts
@@ -4,6 +4,12 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
+import {
+  ForceConfigGet,
+  ForceOrgDisplay,
+  OrgInfo
+} from '@salesforce/salesforcedx-utils-vscode/out/src/cli';
+import { RequestService } from '@salesforce/salesforcedx-utils-vscode/out/src/requestService';
 
 import * as AsyncLock from 'async-lock';
 import { expect } from 'chai';
@@ -29,10 +35,6 @@ import {
   LineBreakpointsInTyperef
 } from '../../../src/breakpoints/lineBreakpoint';
 import {
-  ForceConfigGet,
-  ForceOrgDisplay,
-  OrgInfo,
-  RequestService,
   RunCommand,
   StateCommand,
   StepIntoCommand,
