+++ packages/salesforcedx-apex-debugger/test/unit/commands/baseDebuggerCommand.test.ts
@@ -5,12 +5,13 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
+import { RequestService } from '@salesforce/salesforcedx-utils-vscode/out/src/requestService';
 import { expect } from 'chai';
 import { XHROptions, XHRResponse } from 'request-light';
 import * as sinon from 'sinon';
 import { BaseDebuggerCommand } from '../../../src/commands/baseDebuggerCommand';
 import { DebuggerRequest } from '../../../src/commands/protocol';
-import { RequestService } from '../../../src/commands/requestService';
+
 import {
   CLIENT_ID,
   DEFAULT_CONNECTION_TIMEOUT_MS
