+++ packages/salesforcedx-apex-debugger/test/unit/commands/stepCommands.test.ts
@@ -5,11 +5,11 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
+import { RequestService } from '@salesforce/salesforcedx-utils-vscode/out/src/requestService';
 import { expect } from 'chai';
 import { XHROptions, XHRResponse } from 'request-light';
 import * as sinon from 'sinon';
 import {
-  RequestService,
   StepIntoCommand,
   StepOutCommand,
   StepOverCommand
