+++ packages/salesforcedx-apex-debugger/test/unit/core/streamingClient.test.ts
@@ -5,10 +5,10 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
+import { RequestService } from '@salesforce/salesforcedx-utils-vscode/out/src/requestService';
 import { expect } from 'chai';
 import { Client as FayeClient } from 'faye';
 import * as sinon from 'sinon';
-import { RequestService } from '../../../src/commands';
 import { StreamingClient, StreamingClientInfoBuilder } from '../../../src/core';
 
 describe('Debugger streaming client', () => {
