+++ packages/salesforcedx-apex-debugger/test/unit/core/streamingService.test.ts
@@ -5,9 +5,10 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
+import { OrgInfo } from '@salesforce/salesforcedx-utils-vscode/out/src/cli';
+import { RequestService } from '@salesforce/salesforcedx-utils-vscode/out/src/requestService';
 import { expect } from 'chai';
 import * as sinon from 'sinon';
-import { OrgInfo } from '../../../src/commands/forceOrgDisplay';
 import {
   ApexDebuggerEventType,
   StreamingClient,
@@ -15,7 +16,6 @@ import {
   StreamingService
 } from '../../../src/core';
 import childProcess = require('child_process');
-import { RequestService } from '../../../src/commands';
 
 describe('Debugger streaming service', () => {
   const mockSpawn = require('mock-spawn');
