+++ packages/salesforcedx-utils-vscode/package.json
@@ -17,6 +17,7 @@
     "@types/mocha": "^2.2.38",
     "@types/node": "^6.0.40",
     "@types/shelljs": "^0.7.4",
+    "@types/sinon": "^2.3.2",
     "chai": "^4.0.2",
     "decache": "^4.1.0",
     "glob": "^7.1.2",
@@ -24,9 +25,12 @@
     "mocha": "^3.2.0",
     "mocha-junit-reporter": "^1.13.0",
     "mocha-multi-reporters": "^1.1.4",
+    "mock-spawn": "0.2.6",
     "nyc": "^11.0.2",
     "remap-istanbul": "^0.9.5",
+    "request-light": "0.2.1",
     "shelljs": "^0.7.8",
+    "sinon": "^2.3.6",
     "source-map-support": "^0.4.15",
     "typescript": "2.6.2"
   },
