+++ packages/salesforcedx-utils-vscode/src/cli/index.ts
@@ -22,4 +22,6 @@ export {
   CompositeCliCommandExecutor
 } from './commandExecutor';
 export { CommandOutput } from './commandOutput';
+export { ForceConfigGet } from './forceConfigGet';
+export { ForceOrgDisplay, OrgInfo } from './forceOrgDisplay';
 export { LocalCommandExecution } from './localCommandExecutor';
