+++ packages/salesforcedx-utils-vscode/src/requestService/index.ts
@@ -5,6 +5,5 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-export const CLIENT_ID = 'sfdx-vscode';
-export const DEFAULT_CONNECTION_TIMEOUT_MS = 20000;
-export const SFDX_PROJECT_FILE = 'sfdx-project.json';
+export { BaseCommand } from './baseCommand';
+export { RequestService, RestHttpMethodEnum } from './requestService';
