+++ packages/salesforcedx-vscode-replay-debugger/package.json
@@ -23,7 +23,6 @@
   "categories": ["Debuggers"],
   "dependencies": {
     "@salesforce/salesforcedx-utils-vscode": "42.11.0",
-    "@salesforce/salesforcedx-apex-debugger": "42.11.0",
     "request-light": "0.2.1",
     "vscode-debugadapter": "1.25.0",
     "vscode-debugprotocol": "1.25.0",
