+++ packages/salesforcedx-vscode-replay-debugger/src/commands/queryExistingOverlayActionIdsCommand.ts
@@ -5,7 +5,7 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import { BaseCommand } from '@salesforce/salesforcedx-apex-debugger/out/src/commands';
+import { BaseCommand } from '@salesforce/salesforcedx-utils-vscode/out/src/requestService';
 
 export interface QueryOverlayActionIdsSuccessResult {
   size: number;
