+++ packages/salesforcedx-vscode-replay-debugger/test/unit/commands/apexExecutionOverlayActionCommand.test.ts
@@ -1,11 +1,11 @@
 import {
-  RequestService,
-  RestHttpMethodEnum
-} from '@salesforce/salesforcedx-apex-debugger/out/src/commands';
-import {
   CLIENT_ID,
   DEFAULT_CONNECTION_TIMEOUT_MS
-} from '@salesforce/salesforcedx-apex-debugger/out/src/constants';
+} from '@salesforce/salesforcedx-utils-vscode/out/src/constants';
+import {
+  RequestService,
+  RestHttpMethodEnum
+} from '@salesforce/salesforcedx-utils-vscode/out/src/requestService';
 import { expect } from 'chai';
 import { XHROptions, XHRResponse } from 'request-light';
 import * as sinon from 'sinon';
