+++ packages/salesforcedx-vscode-replay-debugger/test/unit/commands/queryExistingOverlayActionIdsCommand.test.ts
@@ -1,7 +1,7 @@
 import {
   RequestService,
   RestHttpMethodEnum
-} from '@salesforce/salesforcedx-apex-debugger/out/src/commands';
+} from '@salesforce/salesforcedx-utils-vscode/out/src/requestService';
 import { expect } from 'chai';
 import { XHROptions, XHRResponse } from 'request-light';
 import * as sinon from 'sinon';
