+++ packages/salesforcedx-vscode-replay-debugger/tsconfig.json
@@ -2,9 +2,7 @@
   "compilerOptions": {
     "module": "commonjs",
     "target": "es6",
-    "lib": [
-      "es2016"
-    ],
+    "lib": ["es2016", "dom"],
     "sourceMap": true,
     "declaration": true,
     "moduleResolution": "node",
@@ -14,9 +12,5 @@
     "preserveConstEnums": true,
     "strict": true
   },
-  "exclude": [
-    "node_modules",
-    ".vscode-test",
-    "out"
-  ]
+  "exclude": ["node_modules", ".vscode-test", "out"]
 }
